% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{CancelOperationRequest}
\alias{CancelOperationRequest}
\title{Google Cloud RuntimeConfig API Objects 
Provides capabilities for dynamic configuration and coordination for applications running on Google Cloud Platform.}
\usage{
CancelOperationRequest()
}
\value{
CancelOperationRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:10:58
filename: /Users/mark/dev/R/autoGoogleAPI/googleruntimeconfigv1.auto/R/runtimeconfig_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CancelOperationRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for Operations.CancelOperation.
}
\seealso{
Other CancelOperationRequest functions: \code{\link{operations.cancel}}
}
