% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_functions.R
\docType{package}
\name{runtimeconfig_googleAuthR}
\alias{runtimeconfig_googleAuthR}
\alias{runtimeconfig_googleAuthR-package}
\title{Google Cloud RuntimeConfig API
Provides capabilities for dynamic configuration and coordination for applications running on Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:10:58
filename: /Users/mark/dev/R/autoGoogleAPI/googleruntimeconfigv1.auto/R/runtimeconfig_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloudruntimeconfig
\item https://www.googleapis.com/auth/cloud-platform
}
}
